Made comments resolver root query
