+++ eq-author-api/schema/resolvers/base.js
@@ -82,6 +82,8 @@ const {
   getQuestionnaireMetaById,
   createComments,
   saveModel,
+  saveComments,
+  getCommentsForQuestionnaire,
 } = require("../../utils/datastore");
 
 const { QuestionnaireModel } = require("../../db/models/DynamoDB");
@@ -148,6 +150,13 @@ const createNewQuestionnaire = input => {
   };
 };
 
+const publishCommentUpdates = (questionnaire, componentId) => {
+  pubsub.publish("commentsUpdated", {
+    questionnaire,
+    componentId,
+  });
+};
+
 const Resolvers = {
   Query: {
     questionnaires: async (root, args, ctx) => {
@@ -186,6 +195,13 @@ const Resolvers = {
     },
     me: (root, args, ctx) => ctx.user,
     users: () => listUsers(),
+    comments: async (root, { id }, ctx) => {
+      const questionnaireId = ctx.questionnaire.id;
+      const questionnareComments = await getCommentsForQuestionnaire(
+        questionnaireId
+      );
+      return questionnareComments.comments[id] || [];
+    },
   },
 
   Subscription: {
@@ -223,10 +239,12 @@ const Resolvers = {
       subscribe: () => pubsub.asyncIterator(["publishStatusUpdated"]),
     },
     commentsUpdated: {
-      resolve: async ({ pageId, questionnaire }, args, ctx) => {
+      resolve: async ({ componentId, questionnaire }, args, ctx) => {
         ctx.questionnaire = questionnaire;
-        const page = await getPageById(ctx, pageId);
-        return page;
+        const questionnareComments = await getCommentsForQuestionnaire(
+          ctx.questionnaire.id
+        );
+        return questionnareComments.comments[componentId] || [];
       },
       subscribe: () => pubsub.asyncIterator(["commentsUpdated"]),
     },
@@ -701,6 +719,143 @@ const Resolvers = {
       await saveQuestionnaire(ctx.questionnaire);
       return ctx.questionnaire;
     },
+    createComment: async (_, { input }, ctx) => {
+      const { componentId } = input;
+      const questionnaire = ctx.questionnaire;
+      const questionnaireComments = await getCommentsForQuestionnaire(
+        questionnaire.id
+      );
+      const newComment = {
+        id: uuid.v4(),
+        commentText: input.commentText,
+        userId: ctx.user.id,
+        createdTime: new Date(),
+        replies: [],
+      };
+
+      const componentComments = questionnaireComments.comments[componentId];
+
+      if (componentComments) {
+        questionnaireComments.comments[componentId].push(newComment);
+      } else {
+        questionnaireComments.comments[componentId] = [newComment];
+      }
+
+      await saveComments(questionnaireComments);
+      publishCommentUpdates(questionnaire, componentId);
+
+      return newComment;
+    },
+    deleteComment: async (_, { input }, ctx) => {
+      const { componentId } = input;
+      const questionnaire = ctx.questionnaire;
+      const questionnaireComments = await getCommentsForQuestionnaire(
+        questionnaire.id
+      );
+
+      const componentComments = questionnaireComments.comments[componentId];
+      if (componentComments) {
+        remove(componentComments, ({ id }) => id === input.commentId);
+        await saveComments(questionnaireComments);
+      }
+      publishCommentUpdates(questionnaire, componentId);
+      return componentComments;
+    },
+    updateComment: async (_, { input }, ctx) => {
+      const { componentId } = input;
+      const questionnaire = ctx.questionnaire;
+      const questionnaireComments = await getCommentsForQuestionnaire(
+        questionnaire.id
+      );
+      const pageComments = questionnaireComments.comments[componentId];
+
+      if (!pageComments) {
+        throw new Error("No comments found");
+      }
+
+      const commentToEdit = pageComments.find(
+        ({ id }) => id === input.commentId
+      );
+      commentToEdit.commentText = input.commentText;
+      commentToEdit.editedTime = new Date();
+      await saveComments(questionnaireComments);
+
+      publishCommentUpdates(questionnaire, componentId);
+
+      return commentToEdit;
+    },
+    createReply: async (_, { input }, ctx) => {
+      const { componentId } = input;
+      const questionnaire = ctx.questionnaire;
+      const questionnaireComments = await getCommentsForQuestionnaire(
+        questionnaire.id
+      );
+
+      const newReply = {
+        id: uuid.v4(),
+        parentCommentId: input.commentId,
+        commentText: input.commentText,
+        userId: ctx.user.id,
+        createdTime: new Date(),
+      };
+
+      let parentComment = questionnaireComments.comments[componentId].find(
+        ({ id }) => id === input.commentId
+      );
+
+      if (parentComment) {
+        parentComment.replies.push(newReply);
+      } else {
+        parentComment = [newReply];
+      }
+
+      await saveComments(questionnaireComments);
+
+      publishCommentUpdates(questionnaire, componentId);
+
+      return newReply;
+    },
+    updateReply: async (_, { input }, ctx) => {
+      const { componentId } = input;
+      const questionnaire = ctx.questionnaire;
+      const questionnaireComments = await getCommentsForQuestionnaire(
+        questionnaire.id
+      );
+      const replies = questionnaireComments.comments[componentId].find(
+        ({ id }) => id === input.commentId
+      ).replies;
+
+      if (!replies) {
+        throw new Error("No replies found!");
+      }
+
+      const replyToEdit = replies.find(({ id }) => id === input.replyId);
+      replyToEdit.commentText = input.commentText;
+      replyToEdit.editedTime = new Date();
+      await saveComments(questionnaireComments);
+
+      publishCommentUpdates(questionnaire, componentId);
+      return replyToEdit;
+    },
+    deleteReply: async (_, { input }, ctx) => {
+      const { componentId } = input;
+      const questionnaire = ctx.questionnaire;
+      const questionnaireComments = await getCommentsForQuestionnaire(
+        questionnaire.id
+      );
+
+      const replies = questionnaireComments.comments[componentId].find(
+        ({ id }) => id === input.commentId
+      ).replies;
+
+      if (replies) {
+        remove(replies, ({ id }) => id === input.replyId);
+        await saveComments(questionnaireComments);
+      }
+      publishCommentUpdates(questionnaire, componentId);
+
+      return replies;
+    },
   },
 
   Questionnaire: {
@@ -730,6 +885,24 @@ const Resolvers = {
 
   User: {
     displayName: user => user.name || user.email,
+  },
+
+  Comment: {
+    user: ({ userId }) => getUserById(userId),
+  },
+
+  Reply: {
+    user: ({ userId }) => getUserById(userId),
+    parentComment: async ({ parentCommentId, pageId }, args, ctx) => {
+      const questionnaire = ctx.questionnaire;
+      const questionnaireComments = await getCommentsForQuestionnaire(
+        questionnaire.id
+      );
+      const parentComment = questionnaireComments.comments[pageId].find(
+        ({ id }) => id === parentCommentId
+      );
+      return parentComment;
+    },
   },
 
   QuestionnaireInfo: {
