+++ eq-author/src/App/introduction/IntroductionLayout.js
@@ -5,14 +5,19 @@ import EditorLayout from "components/EditorLayout";
 
 import Panel from "components/Panel";
 
-const IntroductionLayout = ({ children }) => (
-  <EditorLayout preview title="Questionnaire Introduction">
+const IntroductionLayout = ({ renderPanel, children }) => (
+  <EditorLayout
+    preview
+    title="Questionnaire Introduction"
+    renderPanel={renderPanel}
+  >
     <Panel>{children}</Panel>
   </EditorLayout>
 );
 
 IntroductionLayout.propTypes = {
   children: PropTypes.node.isRequired,
+  renderPanel: PropTypes.func,
 };
 
 export default IntroductionLayout;
