+++ eq-author/src/components/CommentsPanel/commentSubscription.graphql
@@ -1,30 +1,27 @@
-subscription CommentsUpdated($pageId: ID!) {
-  commentsUpdated(pageId: $pageId) {
+subscription CommentsUpdated($componentId: ID!) {
+  commentsUpdated(id: $componentId) {
     id
-    comments {
+    commentText
+    user {
+      id
+      name
+      picture
+      email
+      displayName
+    }
+    createdTime
+    editedTime
+    replies {
       id
       commentText
+      createdTime
+      editedTime
       user {
         id
         name
         picture
         email
         displayName
-      }
-      createdTime
-      editedTime
-      replies {
-        id
-        commentText
-        createdTime
-        editedTime
-        user {
-          id
-          name
-          picture
-          email
-          displayName
-        }
       }
     }
   }
