+++ eq-author/src/components/CommentsPanel/deleteComment.graphql
@@ -1,8 +1,28 @@
 mutation deleteComment($input: DeleteCommentInput!) {
   deleteComment(input: $input) {
     id
-    comments {
+    commentText
+    user {
       id
+      name
+      picture
+      email
+      displayName
+    }
+    createdTime
+    editedTime
+    replies {
+      id
+      commentText
+      createdTime
+      editedTime
+      user {
+        id
+        name
+        picture
+        email
+        displayName
+      }
     }
   }
 }
